requirements.txt
Django
djangorestframework
markdown
django-filter
exif
ExifRead
ffmpeg-python
piexif
future
geographiclib
geopy
pillow
plum-py
sqlparse
gunicorn
numpy
reverse_geocoder
pycountry
python-magic
urllib3
PyExifTool
